♻️  refactor: 성공/에러 응답 클래스 구분 #1
