+++ src/main/java/sopt/org/motivooServer/domain/HealthCheck.java
@@ -2,6 +2,7 @@ package sopt.org.motivooServer.domain;
 
 import static sopt.org.motivooServer.global.common.response.SuccessType.*;
 
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -13,7 +14,7 @@ import sopt.org.motivooServer.global.common.response.ApiResponse;
 public class HealthCheck {
 
 	@GetMapping
-	public ApiResponse<String> healthCheck() {
+	public ResponseEntity<ApiResponse<String>> healthCheck() {
 		return ApiResponse.success(HEALTH_CHECK_SUCCESS, "test success!");
 	}
 }
