+++ src/main/java/sopt/org/motivooServer/domain/mission/exception/MissionException.java
@@ -1,6 +1,6 @@
 package sopt.org.motivooServer.domain.mission.exception;
 
-import sopt.org.motivooServer.global.common.error.ErrorType;
+import sopt.org.motivooServer.global.common.exception.ErrorType;
 import sopt.org.motivooServer.global.common.exception.BusinessException;
 
 public class MissionException extends BusinessException {
