+++ src/main/java/sopt/org/motivooServer/domain/user/exception/UserException.java
@@ -1,6 +1,6 @@
 package sopt.org.motivooServer.domain.user.exception;
 
-import sopt.org.motivooServer.global.common.error.ErrorType;
+import sopt.org.motivooServer.global.common.exception.ErrorType;
 import sopt.org.motivooServer.global.common.exception.BusinessException;
 
 public class UserException extends BusinessException {
