+++ src/main/java/sopt/org/motivooServer/global/common/exception/BusinessException.java
@@ -1,7 +1,6 @@
 package sopt.org.motivooServer.global.common.exception;
 
 import lombok.Getter;
-import sopt.org.motivooServer.global.common.error.ErrorType;
 
 @Getter
 public class BusinessException extends RuntimeException {
